% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{adding}
\alias{adding}
\title{Adding function}
\usage{
adding(a, b)
}
\arguments{
\item{a}{this should be a float or an integer}

\item{b}{}
}
\value{
Return is a float or an integer
}
\description{
add numbers
}
\examples{
adding(1,2)

}
