% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{some sample data}
\format{
1459 X 2 datafrme

\describe{
\item{index}{country names}
\item{column1}{saleprice}
}
}
\usage{
sampledata
}
\description{
some sample data
}
\keyword{datasets}
